Renamed test/sandbox/utils.js to test/sandbox/utils/random.js
