+++ src/client/components/Task/__stories__/tasks.js
@@ -1,6 +1,6 @@
 import _ from 'lodash'
 
-import { faker } from '../../../../../test/sandbox/utils'
+import { faker } from '../../../../../test/sandbox/utils/random'
 
 const OPTIONS = Array(1000)
   .fill()
