+++ test/component/cypress/specs/Companies/AccountManagement/Objective/ObjectiveForm.cy.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { faker } from '../../../../../../sandbox/utils.js'
+import { faker } from '../../../../../../sandbox/utils/random.js'
 
 import ObjectiveForm from '../../../../../../../src/client/modules/Companies/AccountManagement/Objective/ObjectiveForm'
 import { companyFaker } from '../../../../../../functional/cypress/fakers/companies'
