+++ test/component/cypress/specs/Tasks/TaskForm/TaskFormFields.cy.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { capitalize } from 'lodash'
 
-import { faker } from '../../../../../sandbox/utils.js'
+import { faker } from '../../../../../sandbox/utils/random.js'
 
 import {
   assertFieldDate,
