+++ test/functional/cypress/fakers/administrative-areas.js
@@ -1,4 +1,4 @@
-import { faker } from '../../../sandbox/utils'
+import { faker } from '../../../sandbox/utils/random'
 
 import { listFaker } from './utils'
