+++ test/functional/cypress/fakers/dates.js
@@ -1,4 +1,4 @@
-import { faker } from '../../../sandbox/utils'
+import { faker } from '../../../sandbox/utils/random'
 
 import { addDays } from '../../../../src/client/utils/date'
