+++ test/functional/cypress/fakers/dnb-hierarchy.js
@@ -1,4 +1,4 @@
-import { faker } from '../../../sandbox/utils'
+import { faker } from '../../../sandbox/utils/random'
 
 import { listFaker } from './utils'
 import { addressFaker } from './addresses'
