+++ test/functional/cypress/fakers/eyb-leads.js
@@ -1,4 +1,4 @@
-import { faker, jsf } from '../../../sandbox/utils'
+import { faker, jsf } from '../../../sandbox/utils/random'
 
 import apiSchema from '../../../api-schema.json'
