+++ test/functional/cypress/fakers/investment-project-summary.js
@@ -1,6 +1,6 @@
 import { snakeCase } from 'lodash'
 
-import { faker } from '../../../sandbox/utils'
+import { faker } from '../../../sandbox/utils/random'
 
 import { INVESTMENT_PROJECT_STAGES_LIST } from './constants'
