+++ test/functional/cypress/fakers/investor-type.js
@@ -1,4 +1,4 @@
-import { faker } from '../../../sandbox/utils'
+import { faker } from '../../../sandbox/utils/random'
 
 const investorTypes = [
   'Angel syndicate',
