+++ test/functional/cypress/fakers/reminders.js
@@ -1,4 +1,4 @@
-import { faker } from '../../../sandbox/utils'
+import { faker } from '../../../sandbox/utils/random'
 
 import { relativeDateFaker } from './dates'
 import { investmentProjectCodeFaker } from './investment-projects'
