+++ test/functional/cypress/specs/companies/account-management-spec.js
@@ -1,6 +1,6 @@
 import { format } from 'date-fns'
 
-import { faker } from '../../../../sandbox/utils'
+import { faker } from '../../../../sandbox/utils/random'
 
 import { companyFaker } from '../../fakers/companies'
 import { userFaker } from '../../fakers/users'
