+++ test/functional/cypress/specs/tasks/add-task-spec.js
@@ -1,4 +1,4 @@
-import { faker } from '../../../../sandbox/utils'
+import { faker } from '../../../../sandbox/utils/random'
 
 import { investment, interaction } from '../../fixtures'
 import { investments, tasks, dashboard } from '../../../../../src/lib/urls'
